langchain-server/pyproject.toml
[tool.poetry]
name = "langchain-server"
version = "0.1.0"
description = ""
authors = ["mseo39 <a01063294905@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
langchain = "^0.3.0"
langchain-openai = "^0.2.0"
google-generativeai = "^0.8.0"
langchain-community = "^0.3.0"
langchain-experimental = "^0.3.0"
pydantic = "^2.11.0"
pydantic-settings = "^2.2.1"
python-dotenv = "^1.0.0"
httpx = "^0.27.0"
aiohttp = "^3.9.0"
openai = "^1.10.0"
elasticsearch = "8.12.0"
PyMuPDF = "^1.23.0"
sentence-transformers = "^2.2.2"
scikit-learn = "^1.3.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
